插件返回值为版本号

See merge request rapserver/smarthelper!1091

+++ zfsserver/src/main/java/com/zfs/server/vo/PluginVO.java
@@ -17,7 +17,7 @@ public class PluginVO implements Serializable {
     /**
      * 插件的版本id
      */
-    private Integer pluginv;
+    private long pluginv;
     /**
      * 更新应用下载链接地址
      */
